Fix empty chests
So, because of how it was done with the previous storing of items, some
chests would start off 'looted' simply because they mapped to the same
key in the object for it.

Now, the chest's will create a key using the position as well as the map
name, which should make theoretically unique chests (assuming we can't
have multiple chests in the same spot on the same map).