Enhancement/ability score dialog (#75)
* Add an Ability Score Dialog Before Game Start

This adds a lot of things, so a breakdown here:

  - Some new actions have been added, these are used for
  incrementing/decrementing the ability scores set from the dialog.
  - A new dialog has been added that displays these ability scores, with
  some starting values, and some amount of points. The user is free to
  select how these values are distributed (cannot go below 0, nor above
  20).

* Refactor Ability Score File Layout

* Modularise Abilities

* Change Ability Score Dialog Arrows

* Add Charisma Ability

* Fix Text Sizing in Ability Score Values

* Add Default Ability Score Constant

* Modify Ability Score Dialog

Hopefully this will enable the ability score dialog to be modularised
somewhat, and actually used for level ups (or however we decide to
allocate ability scores) later.

This also removes a graphical nuisance. Originally anytime the state in
the stats reducer would change, the stats bar would flash grey. Moving
the abilities for the dialog into the dialog reducer removes this,
instead opting to flash the stats bar when we actually modify the stats
completely for the player (once they click 'Confirm').

* Minor Ability Dialog UI Tweaks

* Modify Order of Ability Dialog to Match Stats

* Minor Ability Score Dialog UI Tweaks