Add equip button to snackbar closes #64 (#93)
* Add separate styling for new item text

* Access item from payload and attempt to equip

This change allows the snackbar to view the full item from the payload
in the state. The render method displays differently styled text for
when a new item is received as well as an equip button. The button does
not yet successfully equip the item.

* Remove requirement for button title

* Add stats to props

In an attempt to get the dispatch from equipItem to reach the reducer,
stats have been added to the props for snackbar.

* Revert "Remove requirement for button title"

This reverts commit 3332f574a9cfa03269a8ff73cb4769c48aa99a00.

* Style snackbar with equip button

This also changes the new item message from 'GET NEW ITEM' to simply
'NEW ITEM' to save on screen space.

* Alter spacing for sideview

* Remove equip button when receiving potion

* Fix bug where item is not properly equipped

Previously the item would be equipped raw, meaning the item added to the
inventory is not actually equipped but rather a different instance is.
Now the inventory item is equipped.

* Refactor accessing last item in inventory