Fix Issues in Deployment with Vite (#350)
* Configure base path for vite

* Add helper scripts for debugging prod

Leveraging `vite preview` allows us to simulate the production
environment on our local machines, which makes debugging certain issues
a lot easier.

* Fix issue with tile loading

The tiles were attempting to work off of the root path, instead of the
subpath, and therefore failing to load. With this change all of the
tiles now correctly load from the `roll-for-reaction/tiles/` path.

* Refactor GET_REF_URL

Also updates the build target to use ESNext, in an attempt to fix issues
with the Kubernetes deployment.